Removed using of basic retreating functionality (doesn't seems to work/help).
Added fixes for A* to make it faster (hopefully) and actually find the shortest path (for sure).

Separated functions for taking an Ant off getting food or exploring to get food which is close to it because they have different functionality.

Added functionality to interrupt a A* search and route the ant to the current destination if we run out of time (limited to 550 steps at this point).

Fixed timer to work in Linux and Windows without changing include.

Added data for where we've explored bug haven't implemented a use for it

Fixed warnings in visual studio

TODO:
Ant Groups / Flocking
Exploring unexplored (grid[][].isExplored) areas

V2