Changed output of location to cols, rows or x, y
Changed exploreDistance and minExploreDistance to be members of state and are calculated  in relation to the map size and the number of players.

Changes to defending hill code. If no hills are at risk defending ants should be reset to what they were doing.

Added code to kill hills close to ants.

V5